RUB-270 fix storage cleanup review gaps
